+++ src/email/MailChimp/DuplicateEmails.php
@@ -14,6 +14,7 @@ require '../../Include/Functions.php';
 
 use EcclesiaCRM\Utils\InputUtils;
 use EcclesiaCRM\dto\SystemURLs;
+use EcclesiaCRM\dto\SystemConfig;
 
 
 //Set the page title
@@ -23,13 +24,13 @@ require '../../Include/Header.php'; ?>
 
 <div class="box box-body">
 <div class="box-header  with-border">
-  <h3 class="box-title"><?= _("Duplicate Emails")?></h3><div style="float:right"><a href="https://mailchimp.com/en/"><img src="<?= SystemURLs::getRootPath() ?>/Images/Mailchimp_Logo-Horizontal_Black.png" height=25/></a></div>
+  <h3 class="box-title"><?= _("Duplicate Emails")?></h3><div style="float:right"><a href="https://mailchimp.com/<?= substr(SystemConfig::getValue('sLanguage'),0,2) ?>/"><img src="<?= SystemURLs::getRootPath() ?>/Images/Mailchimp_Logo-Horizontal_Black.png" height=25/></a></div>
 </div>
 <table class="table table-striped table-bordered" id="duplicateTable" cellpadding="5" cellspacing="0"  width="100%"></table>
 
 </div>
 
-<script src="<?= SystemURLs::getRootPath() ?>/skin/js/DuplicateEmails.js" ></script>
+<script src="<?= SystemURLs::getRootPath() ?>/skin/js/email/MailChimp/DuplicateEmails.js" ></script>
 
 <?php
 require '../../Include/Footer.php';
